+++ src/features/common/components/icons/DeleteIcon.tsx
@@ -1,7 +1,7 @@
-export const SearchIcon = () => {
+const DeleteIcon = () => {
   return (
     <svg
-      className="w-7 h-16 text-rhyth-blue"
+      className="w-full h-full"
       aria-hidden="true"
       xmlns="http://www.w3.org/2000/svg"
       width="24"
@@ -12,9 +12,12 @@ export const SearchIcon = () => {
       <path
         stroke="currentColor"
         strokeLinecap="round"
+        strokeLinejoin="round"
         strokeWidth="2"
-        d="m21 21-3.5-3.5M17 10a7 7 0 1 1-14 0 7 7 0 0 1 14 0Z"
+        d="M5 7h14m-9 3v8m4-8v8M10 3h4a1 1 0 0 1 1 1v3H9V4a1 1 0 0 1 1-1ZM6 7h12v13a1 1 0 0 1-1 1H7a1 1 0 0 1-1-1V7Z"
       />
     </svg>
   );
 };
+
+export { DeleteIcon };
