+++ src/features/common/components/icons/ExitIcon.tsx
@@ -1,0 +1,21 @@
+const ExitIcon = () => {
+  return (
+    <svg
+      className="w-full h-full"
+      aria-hidden="true"
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+    >
+      <path
+        stroke="currentColor"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="2"
+        d="M16 12H4m12 0-4 4m4-4-4-4m3-4h2a3 3 0 0 1 3 3v10a3 3 0 0 1-3 3h-2"
+      />
+    </svg>
+  );
+};
+
+export { ExitIcon };
