--- src/features/common/components/icons/header/RankingIcon.tsx
@@ -1,36 +1,0 @@
-export const RankingIcon = () => {
-  return (
-    <svg
-      version="1.1"
-      id="_x32_"
-      xmlns="http://www.w3.org/2000/svg"
-      xmlnsXlink="http://www.w3.org/1999/xlink"
-      x="0px"
-      y="0px"
-      viewBox="0 0 512 512"
-      xmlSpace="preserve"
-      className="w-6 h-16"
-    >
-      <g>
-        <title>rhythmate ranking icon</title>
-        <path
-          className="st0 fill-rhyth-orange"
-          d="M512,148.71c0-24.609-19.953-44.563-44.563-44.563c-24.625,0-44.563,19.953-44.563,44.563
-		c0,11.328,4.266,21.641,11.234,29.5c-8.203,18.719-26.391,51.484-52.016,46.359c-32.813-6.578-36.281-60.234-32.719-102.656
-		c17.031-5.578,29.719-21.109,30.688-40.063c1.25-24.578-17.656-45.516-42.25-46.766c-24.563-1.266-45.5,17.656-46.766,42.219
-		c-0.734,14.609,5.672,27.891,16.125,36.563c-13.297,74.219-28.234,90.297-51.172,91.734c-22.938-1.438-37.875-17.516-51.172-91.734
-		c10.453-8.672,16.859-21.953,16.125-36.563c-1.266-24.563-22.203-43.484-46.766-42.219c-24.594,1.25-43.5,22.188-42.25,46.766
-		c0.969,18.953,13.656,34.484,30.688,40.063c3.563,42.422,0.094,96.078-32.719,102.656c-25.625,5.109-43.813-27.641-52-46.344
-		c6.969-7.875,11.219-18.188,11.219-29.516c0-24.609-19.938-44.563-44.563-44.563C19.953,104.147,0,124.101,0,148.71
-		c0,20.328,13.625,37.438,32.219,42.797l42.688,163.906h172.563h17.063h172.563l42.688-163.906
-		C498.375,186.147,512,169.038,512,148.71z"
-        />
-        <path
-          className="st0 fill-rhyth-orange"
-          d="M423.813,401.132H88.188c-16.75,0-30.344,13.578-30.344,30.344v15.172c0,16.75,13.594,30.328,30.344,30.328
-		h335.625c16.75,0,30.344-13.578,30.344-30.328v-15.172C454.156,414.71,440.563,401.132,423.813,401.132z"
-        />
-      </g>
-    </svg>
-  );
-};
