--- src/features/common/components/icons/header/index.ts
@@ -1,5 +1,0 @@
-export * from "./BadgeIcon";
-export * from "./BellIcon";
-export * from "./RankingIcon";
-export * from "./SearchIcon";
-export * from "./TagIcon";
