+++ src/features/common/components/icons/index.ts
@@ -1,4 +1,6 @@
 export * from "./AnalyticsIcon";
+export * from "./BadgeIcon";
+export * from "./DeleteIcon";
 export * from "./ManageIcon";
 export * from "./QuestsIcon";
 export * from "./ProfileIcon";
@@ -7,6 +9,7 @@ export * from "./HonoIcon";
 export * from "./TagIcon";
 export * from "./PencilIcon";
 export * from "./EditIcon";
+export * from "./ExitIcon";
 export * from "./ChevronUpIcon";
 export * from "./ChevronDownIcon";
 export * from "./ChevronLeftIcon";
@@ -14,6 +17,8 @@ export * from "./ChevronRightIcon";
 export * from "./CloseIcon";
 export * from "./StarIcon";
 export * from "./SwitchIcon";
+export * from "./SearchIcon";
 export * from "./AddIcon";
 export * from "./ChatIcon";
 export * from "./ChatIcon2";
+export * from "./LoadingIcon";
