+++ src/features/tags/components/TagsDeleteButton.tsx
@@ -1,4 +1,5 @@
 import type { FC } from "react";
+import { DeleteIcon } from "../../common/components";
 
 type Props = {
   onClick: () => void;
@@ -7,23 +8,9 @@ type Props = {
 export const TagsDeleteButton: FC<Props> = ({ onClick }) => {
   return (
     <button type="button" onClick={onClick}>
-      <svg
-        className="w-5 h-5 text-rhyth-red font-noto-sans font-bold"
-        aria-hidden="true"
-        xmlns="http://www.w3.org/2000/svg"
-        width="24"
-        height="24"
-        fill="none"
-        viewBox="0 0 24 24"
-      >
-        <path
-          stroke="currentColor"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth="2"
-          d="M5 7h14m-9 3v8m4-8v8M10 3h4a1 1 0 0 1 1 1v3H9V4a1 1 0 0 1 1-1ZM6 7h12v13a1 1 0 0 1-1 1H7a1 1 0 0 1-1-1V7Z"
-        />
-      </svg>
+      <div className="w-5 h-5 text-rhyth-red">
+        <DeleteIcon />
+      </div>
     </button>
   );
 };
